[mk][top] Move lib/debug.c to top/debug.cc
