+++ kernel/rules.mk
@@ -11,7 +11,6 @@ MODULE := $(LOCAL_DIR)
 
 MODULE_DEPS := \
 	lib/libc \
-	lib/debug \
 	lib/heap \
 	lib/ktl
