--- lib/debug/rules.mk
@@ -1,8 +1,0 @@
-LOCAL_DIR := $(GET_LOCAL_DIR)
-
-MODULE := $(LOCAL_DIR)
-
-MODULE_SRCS += \
-	$(LOCAL_DIR)/debug.c
-
-include make/module.mk
