+++ lib/debugcommands/rules.mk
@@ -3,7 +3,6 @@ LOCAL_DIR := $(GET_LOCAL_DIR)
 MODULE := $(LOCAL_DIR)
 
 MODULE_DEPS += \
-	lib/debug \
 	lib/console
 
 MODULE_SRCS += \
