+++ top/debug.cc
@@ -8,10 +8,54 @@
 
 #include "debug.h"
 
-#include <lk/debug.h>
+#include <align.h>
+#include <ctype.h>
 #include <platform.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <zircon/types.h>
 
 #include <__verbose_abort>
+
+#include <arch/ops.h>
+#include <kernel/spinlock.h>
+#include <ktl/algorithm.h>
+#include <platform/debug.h>
+
+// Note <ktl/enforce.h> can't be used here because we need to define a function
+// that was declared in the std:: namespace by <__verbose_abort>.
+
+static ssize_t __panic_stdio_read(io_handle_t* io, char* s, size_t len) {
+  if (len == 0)
+    return 0;
+
+  int err = platform_pgetc(s, false);
+  if (err < 0)
+    return err;
+
+  return 1;
+}
+
+static ssize_t __panic_stdio_write(io_handle_t* io, const char* s, size_t len) {
+  for (size_t i = 0; i < len; i++) {
+    platform_pputc(s[i]);
+  }
+  return len;
+}
+
+static const io_handle_hooks_t panic_hooks = {
+    .write = __panic_stdio_write,
+    .read = __panic_stdio_read,
+};
+
+static io_handle_t panic_io = {
+    .magic = IO_HANDLE_MAGIC,
+    .hooks = &panic_hooks,
+};
+
+FILE stdout_panic_buffer = {
+    .io = &panic_io,
+};
 
 namespace {
 
@@ -26,7 +70,7 @@ namespace {
 __ALWAYS_INLINE inline void PanicStart(void* pc, void* frame) {
   // platform_panic_start();
 
-  fprintf(get_panic_fd(),
+  fprintf(&stdout_panic_buffer,
           "\n"
           "*** KERNEL PANIC (caller pc: %p, stack frame: %p):\n"
           "*** ",
@@ -42,7 +86,7 @@ __ALWAYS_INLINE inline void PanicStart(void* pc, void* frame) {
 // appearing in the backtrace.
 __ALWAYS_INLINE __NO_RETURN inline void PanicFinish() {
   // Add a newline between the panic message and the stack trace.
-  fprintf(get_panic_fd(), "\n");
+  fprintf(&stdout_panic_buffer, "\n");
 
   // platform_halt(HALT_ACTION_HALT, ZirconCrashReason::Panic);
   platform_halt(HALT_ACTION_HALT, HALT_REASON_SW_PANIC);
@@ -58,12 +102,12 @@ bool EndsWith(const char* str, char x) {
   PanicStart(pc, frame);
 
   // Print the user message.
-  vfprintf(get_panic_fd(), fmt, ap);
+  vfprintf(&stdout_panic_buffer, fmt, ap);
   va_end(ap);
 
   // Add a newline to the end of the panic message if it was missing.
   if (!EndsWith(fmt, '\n')) {
-    fprintf(get_panic_fd(), "\n");
+    fprintf(&stdout_panic_buffer, "\n");
   }
 
   PanicFinish();
@@ -71,9 +115,46 @@ bool EndsWith(const char* str, char x) {
 
 }  // namespace
 
+void spin(uint32_t usecs) {
+  lk_bigtime_t start = current_time_hires();
+
+  while ((current_time_hires() - start) < usecs)
+    ;
+}
+
+void panic(const char* fmt, ...) {
+  va_list ap;
+  va_start(ap, fmt);
+  vpanic(__GET_CALLER(), __GET_FRAME(), fmt, ap);
+}
+
 // Inline functions in libc++ headers call this.
 [[noreturn]] void std::__libcpp_verbose_abort(const char* format, ...) {
   va_list ap;
   va_start(ap, format);
   vpanic(__GET_CALLER(), __GET_FRAME(), format, ap);
+}
+
+void assert_fail_msg(const char* file, int line, const char* expression, const char* fmt, ...) {
+  PanicStart(__GET_CALLER(), __GET_FRAME());
+
+  // Print the user message.
+  fprintf(&stdout_panic_buffer, "ASSERT FAILED at (%s:%d): %s\n", file, line, expression);
+  va_list ap;
+  va_start(ap, fmt);
+  vfprintf(&stdout_panic_buffer, fmt, ap);
+  va_end(ap);
+
+  // Add a newline to the end of the panic message if it was missing.
+  if (!EndsWith(fmt, '\n')) {
+    fprintf(&stdout_panic_buffer, "\n");
+  }
+
+  PanicFinish();
+}
+
+void assert_fail(const char* file, int line, const char* expression) {
+  PanicStart(__GET_CALLER(), __GET_FRAME());
+  fprintf(&stdout_panic_buffer, "ASSERT FAILED at (%s:%d): %s\n", file, line, expression);
+  PanicFinish();
 }
